requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
sqlalchemy
openai
python-dotenv
